pull전 커밋
